Refactored test to reflect file structure
