+++ gpslogger/src/test/java/com/mendhak/gpslogger/loggers/Gpx10WriteHandlerTest.java
@@ -1,117 +1,21 @@
 package com.mendhak.gpslogger.loggers;
 
 import android.location.Location;
-import android.os.Bundle;
-
 import android.test.suitebuilder.annotation.SmallTest;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.runners.MockitoJUnitRunner;
 
-import static org.hamcrest.CoreMatchers.*;
+import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.mockito.Matchers.eq;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 
 @SmallTest
 @RunWith(MockitoJUnitRunner.class)
-public class FileLoggerTests  {
-
-
-    private static final class MockLocations {
-
-        private final Location loc;
-        private final Bundle bundle;
-
-        private MockLocations(Location loc, Bundle bundle) {
-            this.loc = loc;
-            this.bundle = bundle;
-        }
-
-        public static MockLocations builder(String providerName, double lat, double lon){
-
-            MockLocations m = new MockLocations(mock(Location.class), mock(Bundle.class));
-            when(m.loc.getProvider()).thenReturn(providerName);
-            when(m.loc.getLatitude()).thenReturn(lat);
-            when(m.loc.getLongitude()).thenReturn(lon);
-            return m;
-        }
-
-
-        public MockLocations withAccuracy(float val) {
-            when(loc.hasAccuracy()).thenReturn(true);
-            when(loc.getAccuracy()).thenReturn(val);
-            return this;
-        }
-
-        public Location build(){
-            return loc;
-        }
-
-        public MockLocations withAltitude(double altitude) {
-            when(loc.hasAltitude()).thenReturn(true);
-            when(loc.getAltitude()).thenReturn(altitude);
-            return this;
-        }
-
-        public MockLocations withBearing(float bearing) {
-            when(loc.hasBearing()).thenReturn(true);
-            when(loc.getBearing()).thenReturn(bearing);
-            return this;
-        }
-
-
-        public MockLocations withSpeed(float speed) {
-            when(loc.hasSpeed()).thenReturn(true);
-            when(loc.getSpeed()).thenReturn(speed);
-            return this;
-        }
-
-        public MockLocations putExtra(String k, String v) {
-
-//            Bundle b = new Bundle();
-//            b.putString("HDOP", "LOOKATTHISHDOP!");
-
-            when(loc.getExtras()).thenReturn(bundle);
-            when(bundle.getString(eq(k))).thenReturn(v);
-
-            return this;
-        }
-    }
+public class Gpx10WriteHandlerTest {
 
     @Test
-    public void testWaypointXml_BasicInfo(){
-        Gpx10AnnotateHandler annotateHandler = new Gpx10AnnotateHandler(null, null, null, null);
-
-
-        Location loc = MockLocations.builder("MOCK", 12.193, 19.111).build();
-
-
-        when(loc.hasAccuracy()).thenReturn(false);
-
-        String actual =  annotateHandler.GetWaypointXml(loc, "2011-09-17T18:45:33Z", "This is the annotation");
-        String expected = "\n<wpt lat=\"12.193\" lon=\"19.111\"><time>2011-09-17T18:45:33Z</time><name>This is the annotation</name><src>MOCK</src></wpt>\n";
-
-        assertThat("Basic waypoint XML", actual, is(expected));
-    }
-
-
-    @Test
-    public void testWaypointXml_WithAltitude(){
-        Gpx10AnnotateHandler annotateHandler = new Gpx10AnnotateHandler(null, null, null, null);
-
-        Location loc = MockLocations.builder("MOCK", 12.193, 19.111).withAltitude(9001d).build();
-
-        String actual =  annotateHandler.GetWaypointXml(loc, "2011-09-17T18:45:33Z", "This is the annotation");
-        String expected = "\n<wpt lat=\"12.193\" lon=\"19.111\"><ele>9001.0</ele><time>2011-09-17T18:45:33Z</time><name>This is the annotation</name><src>MOCK</src></wpt>\n";
-
-        assertThat("Basic waypoint XML", actual, is(expected));
-    }
-
-    @Test
-    public void testTrackPointXml_LatLongOnly(){
+    public void GetTrackpointXml_BasicLocation_BasicTrkptNodeReturned(){
 
         Gpx10WriteHandler writeHandler = new Gpx10WriteHandler(null, null, null, false, 41);
 
@@ -125,7 +29,7 @@ public class FileLoggerTests  {
 
 
     @Test
-    public void testTrackPointXml_ExtraInfo(){
+    public void GetTrackPointXml_LocationWithAltBearingSpeed_TrkptWithEleCourseSpeedReturned(){
 
         Gpx10WriteHandler writeHandler = new Gpx10WriteHandler(null, null, null, false, 41);
 
@@ -140,7 +44,7 @@ public class FileLoggerTests  {
 
 
     @Test
-    public void testTrackPointXml_ExtraInfoWithoutSatellites(){
+    public void GetTrackPointXml_LocationWithoutSatellites_TrkptNodeReturned(){
 
         Gpx10WriteHandler writeHandler = new Gpx10WriteHandler(null, null, null, false, 0);
 
@@ -159,7 +63,7 @@ public class FileLoggerTests  {
     }
 
     @Test
-    public void testTrackPointXml_NewTrackSegment(){
+    public void GetTrackPointXml_NewTrackSegmentPref_NewTrkSegReturned(){
 
         Gpx10WriteHandler writeHandler = new Gpx10WriteHandler(null, null, null, true, 0);
 
@@ -200,7 +104,7 @@ public class FileLoggerTests  {
 
 
     @Test
-    public void testTrackPointXml_BundledGeoIdHeight(){
+    public void GetTrackPointXml_BundledGeoIdHeight_GeoIdHeightNode(){
 
         Gpx10WriteHandler writeHandler = new Gpx10WriteHandler(null, null, null, true, 0);
 
@@ -217,22 +121,4 @@ public class FileLoggerTests  {
 
         assertThat("Trackpoint XML with a geoid height", actual, is(expected));
     }
-
-    @Test
-    public void testPlacemarkXml_BasicInfo() {
-
-        Kml22AnnotateHandler kmlHandler = new Kml22AnnotateHandler(null, null, null);
-        Location loc = MockLocations.builder("MOCK", 12.193,19.111)
-                .withAltitude(9001d)
-                .withBearing(91.88f)
-                .withSpeed(188.44f)
-                .build();
-
-        String actual = kmlHandler.GetPlacemarkXml("This is the annotation",loc);
-        String expected = "<Placemark><name>This is the annotation</name><Point><coordinates>19.111,12.193,9001.0</coordinates></Point></Placemark>\n";
-
-        assertThat("Basic Placemark XML", actual, is(expected));
-    }
-
-
 }
