+++ gpslogger/src/test/java/com/mendhak/gpslogger/loggers/nmea/NmeaSentenceTest.java
@@ -1,8 +1,6 @@
-package com.mendhak.gpslogger;
-
+package com.mendhak.gpslogger.loggers.nmea;
 
 import android.test.suitebuilder.annotation.SmallTest;
-import com.mendhak.gpslogger.loggers.nmea.NmeaSentence;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.runners.MockitoJUnitRunner;
@@ -11,9 +9,10 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 
+
 @SmallTest
 @RunWith(MockitoJUnitRunner.class)
-public class NmeaSentenceTests {
+public class NmeaSentenceTest {
 
     @Test
     public void NmeaSentence_EmptyNmeaSentence_VDOPIsNull(){
@@ -70,6 +69,4 @@ public class NmeaSentenceTests {
         NmeaSentence nmeaSentence = new NmeaSentence(null);
         assertThat("Null NMEA string", nmeaSentence.getLatestHdop(), nullValue());
     }
-
-
 }
